==> notebooks/02_GenerateFirstSetofPolygonsusingDatasetIDswithOceanFiltering.ipynb <==
import logging
import os

import click
import datacube
import fsspec

from deafrica_waterbodies.cli.logs import logging_setup
from deafrica_waterbodies.io import check_dir_exists, check_file_exists, check_if_s3_uri
from deafrica_waterbodies.make_polygons import check_wetness_thresholds, get_polygons_from_dataset_with_land_sea_mask_filtering

import os

# These are the default AWS configurations for the Analysis Sandbox.
# that are set in the environmnet variables.
aws_default_config = {
    # "AWS_NO_SIGN_REQUEST": "YES",
    "AWS_SECRET_ACCESS_KEY": "fake",
    "AWS_ACCESS_KEY_ID": "fake",
}

# To access public bucket, need to remove the AWS credentials in
# the environment variables or the following error will occur.
# PermissionError: The AWS Access Key Id you provided does not exist in our records.

for key in aws_default_config.keys():
    if key in os.environ:
        del os.environ[key]

verbose = 1
primary_threshold: float = 0.1
secondary_threshold: float = 0.05
minimum_valid_observations: int = 128
output_directory = "s3://deafrica-waterbodies-dev/test_out_dir/0-0-1/shapefile2"
dataset_ids_text_file = "s3://deafrica-waterbodies-dev/test_out_dir/0-0-1/dataset_ids.txt"
overwrite = True
land_sea_mask_fp = "data/af_msk_3s.tif"

import xarray as xr
def filter_hydrosheds_land_mask(hydrosheds_land_mask: xr.DataArray) -> xr.DataArray:
    """
    Function to filter the HydroSHEDs Land Mask into a boolean mask.
    """
    # Indicator values: 1 = land, 2 = ocean sink, 3 = inland sink, 255 is no data.
    boolean_mask = (hydrosheds_land_mask != 255) & (hydrosheds_land_mask != 2)
    return boolean_mask

# Set up logger.
logging_setup(verbose=verbose)
_log = logging.getLogger(__name__)

# Support pathlib paths.
output_directory = str(output_directory)
dataset_ids_text_file = str(dataset_ids_text_file)

# Parameters to use when loading datasets.
dask_chunks = {"x": 3200, "y": 3200, "time": 1}
resolution = (-30, 30)
output_crs = "EPSG:6933"

# Read the dataset ids from the text file.
if not check_file_exists(dataset_ids_text_file):
    _log.error(f"Could not find text file {dataset_ids_text_file}!")
    raise FileNotFoundError(f"Could not find text file {dataset_ids_text_file}!")
else:
    if check_if_s3_uri(dataset_ids_text_file):
        fs = fsspec.filesystem("s3")
    else:
        fs = fsspec.filesystem("file")
    with fs.open(dataset_ids_text_file, "r") as file:
        lines = file.readlines()
        dataset_ids = [line.strip() for line in lines]

# Directory to write generated waterbody polygons to.
polygons_from_thresholds_dir = os.path.join(output_directory, "polygons_from_thresholds")

# Set the filesystem to use.
if check_if_s3_uri(polygons_from_thresholds_dir):
    fs = fsspec.filesystem("s3")
else:
    fs = fsspec.filesystem("file")

# Check if the directory exists. If it does not, create it.
if not check_dir_exists(polygons_from_thresholds_dir):
    fs.mkdirs(polygons_from_thresholds_dir, exist_ok=True)
    _log.info(f"Created directory {polygons_from_thresholds_dir}")

# Check if the wetness thresholds have been set correctly.
minimum_wet_thresholds = [secondary_threshold, primary_threshold]
_log.info(check_wetness_thresholds(minimum_wet_thresholds))

# Connect to the datacube.
dc = datacube.Datacube(app="GenerateWaterbodyPolygons")

# For each dataset id, threshold the scene to generate the primary and secondary threshold
# waterbody polygons.
for dataset_id in dataset_ids:
    primary_threshold_polygons_fp = os.path.join(
        polygons_from_thresholds_dir, f"{dataset_id}_primary_threshold_polygons.parquet"
    )
    secondary_threshold_polygons_fp = os.path.join(
        polygons_from_thresholds_dir, f"{dataset_id}_secondary_threshold_polygons.parquet"
    )

    if not overwrite:
        _log.info(
            f"Checking existence of {primary_threshold_polygons_fp} and {secondary_threshold_polygons_fp}"
        )
        exists = check_file_exists(primary_threshold_polygons_fp) and check_file_exists(
            secondary_threshold_polygons_fp
        )

    if overwrite or not exists:
        (
            primary_threshold_polygons,
            secondary_threshold_polygons,
        ) = get_polygons_from_dataset_with_land_sea_mask_filtering(
            dataset_id=dataset_id,
            dask_chunks=dask_chunks,
            resolution=resolution,
            output_crs=output_crs,
            min_valid_observations=minimum_valid_observations,
            primary_threshold=primary_threshold,
            secondary_threshold=secondary_threshold,
            dc=dc,
            land_sea_mask_fp=land_sea_mask_fp,
            filter_land_sea_mask=filter_hydrosheds_land_mask
        )
        # Write the polygons to parquet files.
        primary_threshold_polygons.to_parquet(primary_threshold_polygons_fp)
        secondary_threshold_polygons.to_parquet(secondary_threshold_polygons_fp)
